\name{ABRAID modes}
\alias{abraid_modes}

\docType{data}
\title{
model run modes available to runABRAID
}

\description{
A list of model modes currently supported for the ABRAID system, used only by \code{runABRAID}. This is not intended to be used more widely.
}

\usage{
  data(abraid_modes)
}

\format{
The file \code{data/abraid_modes.txt} is a newline-delimited file with each line giving a separate string indicating a possible \code{runABRAID} mode. There must be a newline at the end of the file.
}

